+++ Runtime/Model/BacktraceData.cs
@@ -211,7 +211,7 @@ namespace Backtrace.Unity.Model
             LangVersion = "Mono";
 #endif
 
-            AgentVersion = "2.1.3";
+            AgentVersion = "2.1.4";
             Classifier = Report.ExceptionTypeReport ? new[] { Report.Classifier } : null;
         }
     }